refactor object setup into modular placements
